Refactor DeleteModal location
