+++ frontend/src/lib/components/DeleteModal.svelte
